Refactoring; writing iterators to transform while-loops into for-loops
